Fixed Input tests Part I
